requirements-linux.txt
httpx
pytest
black
isort
mypy
fastapi
uvicorn
python-multipart
python-dotenv
requests
pymongo
colorlog